- correct extensions
