+++ tdrs-frontend/src/components/FileUploadForms/SectionFileUploadForm.test.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { fireEvent, waitFor, render } from '@testing-library/react'
 import { Provider } from 'react-redux'
 import configureStore from '../../configureStore'
-import QuarterFileUploadForm from './QuarterFileUploadForm'
+import SectionFileUploadForm from './SectionFileUploadForm'
 import { ReportsProvider } from '../Reports/ReportsContext'
-import * as reportsActions from '../../actions/reports'
+import { fileUploadSections } from '../../reducers/reports'
 import { useFormSubmission } from '../../hooks/useFormSubmission'
 import { useEventLogger } from '../../utils/eventLogger'
 
@@ -20,7 +20,14 @@ jest.mock('@uswds/uswds/src/js/components', () => ({
 // Mock FileUpload component
 jest.mock('../FileUpload', () => ({
   __esModule: true,
-  default: ({ section, year, quarter, fileType, label, setLocalAlertState }) => (
+  default: ({
+    section,
+    year,
+    quarter,
+    fileType,
+    label,
+    setLocalAlertState,
+  }) => (
     <div data-testid={`file-upload-${section}`}>
       <label>{label}</label>
       <input
@@ -56,10 +63,7 @@ const initialState = {
 
 const mockStore = (initial = initialState) => configureStore(initial)
 
-const makeTestFile = (name, contents = ['test'], type = 'text/plain') =>
-  new File(contents, name, { type })
-
-describe('QuarterFileUploadForm', () => {
+describe('SectionFileUploadForm', () => {
   let mockExecuteSubmission
   let mockLogger
   let mockDispatch
@@ -88,35 +92,51 @@ describe('QuarterFileUploadForm', () => {
     jest.clearAllMocks()
   })
 
-  const renderComponent = (storeState = initialState, stt = { id: 1 }) => {
+  const renderComponent = (
+    storeState = initialState,
+    stt = { id: 1, num_sections: 4 }
+  ) => {
     const store = mockStore(storeState)
     mockDispatch = jest.spyOn(store, 'dispatch')
 
     return render(
       <Provider store={store}>
         <ReportsProvider>
-          <QuarterFileUploadForm stt={stt} />
+          <SectionFileUploadForm stt={stt} />
         </ReportsProvider>
       </Provider>
     )
   }
 
   describe('Rendering', () => {
-    it('renders all four quarter file upload inputs', () => {
+    it('renders all four section file upload inputs by default', () => {
       const { getByTestId } = renderComponent()
 
+      expect(getByTestId('file-upload-Active Case Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Closed Case Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Aggregate Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Stratum Data')).toBeInTheDocument()
+    })
+
+    it('renders correct number of sections based on stt.num_sections', () => {
+      const stt = { id: 1, num_sections: 2 }
+      const { getByTestId, queryByTestId } = renderComponent(initialState, stt)
+
+      expect(getByTestId('file-upload-Active Case Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Closed Case Data')).toBeInTheDocument()
       expect(
-        getByTestId('file-upload-Quarter 1 (October - December)')
-      ).toBeInTheDocument()
-      expect(
-        getByTestId('file-upload-Quarter 2 (January - March)')
-      ).toBeInTheDocument()
-      expect(
-        getByTestId('file-upload-Quarter 3 (April - June)')
-      ).toBeInTheDocument()
-      expect(
-        getByTestId('file-upload-Quarter 4 (July - September)')
-      ).toBeInTheDocument()
+        queryByTestId('file-upload-Aggregate Data')
+      ).not.toBeInTheDocument()
+      expect(queryByTestId('file-upload-Stratum Data')).not.toBeInTheDocument()
+    })
+
+    it('renders 4 sections when stt is undefined', () => {
+      const { getByTestId } = renderComponent(initialState, undefined)
+
+      expect(getByTestId('file-upload-Active Case Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Closed Case Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Aggregate Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Stratum Data')).toBeInTheDocument()
     })
 
     it('renders submit and cancel buttons', () => {
@@ -138,6 +158,15 @@ describe('QuarterFileUploadForm', () => {
 
       expect(fileInput.init).toHaveBeenCalled()
     })
+
+    it('renders section labels with correct format', () => {
+      const { getByText } = renderComponent()
+
+      expect(getByText('Section 1 TANF - Active Case Data')).toBeInTheDocument()
+      expect(getByText('Section 2 TANF - Closed Case Data')).toBeInTheDocument()
+      expect(getByText('Section 3 TANF - Aggregate Data')).toBeInTheDocument()
+      expect(getByText('Section 4 TANF - Stratum Data')).toBeInTheDocument()
+    })
   })
 
   describe('Form Submission', () => {
@@ -157,7 +186,9 @@ describe('QuarterFileUploadForm', () => {
       await waitFor(() => {
         const alert = getByRole('alert')
         expect(alert).toBeInTheDocument()
-        expect(alert).toHaveTextContent('No changes have been made to data files')
+        expect(alert).toHaveTextContent(
+          'No changes have been made to data files'
+        )
       })
 
       expect(mockExecuteSubmission).not.toHaveBeenCalled()
@@ -166,9 +197,10 @@ describe('QuarterFileUploadForm', () => {
     it('submits successfully with uploaded files', async () => {
       const uploadedFile = {
         fileName: 'test.txt',
-        section: 'Quarter 1 (October - December)',
-        fileType: 'pia',
+        section: 'Active Case Data',
+        fileType: 'tanf',
         year: '2024',
+        quarter: 'Q1',
       }
 
       const storeState = {
@@ -192,26 +224,19 @@ describe('QuarterFileUploadForm', () => {
       })
     })
 
-    it('transforms files correctly before submission', async () => {
-      const uploadedFiles = [
-        {
-          fileName: 'q1.txt',
-          section: 'Quarter 1 (October - December)',
-          fileType: 'pia',
-          year: '2024',
-        },
-        {
-          fileName: 'q2.txt',
-          section: 'Quarter 2 (January - March)',
-          fileType: 'pia',
-          year: '2024',
-        },
-      ]
+    it('passes correct parameters to submit action', async () => {
+      const uploadedFile = {
+        fileName: 'test.txt',
+        section: 'Active Case Data',
+        fileType: 'tanf',
+        year: '2024',
+        quarter: 'Q1',
+      }
 
       const storeState = {
         ...initialState,
         reports: {
-          submittedFiles: uploadedFiles,
+          submittedFiles: [uploadedFile],
         },
       }
 
@@ -225,26 +250,18 @@ describe('QuarterFileUploadForm', () => {
       fireEvent.click(submitButton)
 
       await waitFor(() => {
-        expect(mockExecuteSubmission).toHaveBeenCalled()
+        expect(mockExecuteSubmission).toHaveBeenCalledWith(expect.any(Function))
       })
-
-      // Verify executeSubmission was called with a function
-      expect(mockExecuteSubmission).toHaveBeenCalledWith(expect.any(Function))
     })
 
-    it('formats quarters correctly in success message', async () => {
+    it('formats sections correctly for single file', () => {
       const uploadedFiles = [
         {
-          fileName: 'q1.txt',
-          section: 'Quarter 1 (October - December)',
-          fileType: 'pia',
+          fileName: 'test.txt',
+          section: 'Active Case Data',
+          fileType: 'tanf',
           year: '2024',
-        },
-        {
-          fileName: 'q3.txt',
-          section: 'Quarter 3 (April - June)',
-          fileType: 'pia',
-          year: '2024',
+          quarter: 'Q1',
         },
       ]
 
@@ -255,26 +272,46 @@ describe('QuarterFileUploadForm', () => {
         },
       }
 
-      mockExecuteSubmission.mockImplementation(async (fn) => {
-        await fn()
-      })
+      renderComponent(storeState)
+      // Component renders without errors
+    })
 
-      const { getByText } = renderComponent(storeState)
+    it('formats sections correctly for multiple files with "and"', () => {
+      const uploadedFiles = [
+        {
+          fileName: 'test1.txt',
+          section: 'Active Case Data',
+          fileType: 'tanf',
+          year: '2024',
+          quarter: 'Q1',
+        },
+        {
+          fileName: 'test2.txt',
+          section: 'Closed Case Data',
+          fileType: 'tanf',
+          year: '2024',
+          quarter: 'Q1',
+        },
+      ]
 
-      const submitButton = getByText('Submit Data Files')
-      fireEvent.click(submitButton)
+      const storeState = {
+        ...initialState,
+        reports: {
+          submittedFiles: uploadedFiles,
+        },
+      }
 
-      await waitFor(() => {
-        expect(mockExecuteSubmission).toHaveBeenCalled()
-      })
+      renderComponent(storeState)
+      // Component renders without errors
     })
 
     it('handles submission errors gracefully', async () => {
       const uploadedFile = {
         fileName: 'test.txt',
-        section: 'Quarter 1 (October - December)',
-        fileType: 'pia',
+        section: 'Active Case Data',
+        fileType: 'tanf',
         year: '2024',
+        quarter: 'Q1',
       }
 
       const storeState = {
@@ -286,6 +323,8 @@ describe('QuarterFileUploadForm', () => {
 
       const error = new Error('Submission failed')
       mockExecuteSubmission.mockRejectedValue(error)
+
+      const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation()
 
       const { getByText, getByRole } = renderComponent(storeState)
 
@@ -299,6 +338,13 @@ describe('QuarterFileUploadForm', () => {
           'An error occurred during submission. Please try again.'
         )
       })
+
+      expect(consoleErrorSpy).toHaveBeenCalledWith(
+        'Error during form submission:',
+        error
+      )
+
+      consoleErrorSpy.mockRestore()
     })
 
     it('disables submit button when isSubmitting is true', () => {
@@ -328,9 +374,10 @@ describe('QuarterFileUploadForm', () => {
     it('calls handleOpenFeedbackWidget after successful submission', async () => {
       const uploadedFile = {
         fileName: 'test.txt',
-        section: 'Quarter 1 (October - December)',
-        fileType: 'pia',
+        section: 'Active Case Data',
+        fileType: 'tanf',
         year: '2024',
+        quarter: 'Q1',
       }
 
       const storeState = {
@@ -358,15 +405,76 @@ describe('QuarterFileUploadForm', () => {
         expect(mockDispatch).toHaveBeenCalled()
       })
     })
+
+    it('passes ssp flag correctly for ssp-moe file type', async () => {
+      const uploadedFile = {
+        fileName: 'test.txt',
+        section: 'Active Case Data',
+        fileType: 'ssp-moe',
+        year: '2024',
+        quarter: 'Q1',
+      }
+
+      const storeState = {
+        ...initialState,
+        reports: {
+          submittedFiles: [uploadedFile],
+        },
+      }
+
+      mockExecuteSubmission.mockImplementation(async (fn) => {
+        await fn()
+      })
+
+      const { getByText } = renderComponent(storeState)
+
+      const submitButton = getByText('Submit Data Files')
+      fireEvent.click(submitButton)
+
+      await waitFor(() => {
+        expect(mockExecuteSubmission).toHaveBeenCalled()
+      })
+    })
+
+    it('passes ssp flag as false for non-ssp-moe file types', async () => {
+      const uploadedFile = {
+        fileName: 'test.txt',
+        section: 'Active Case Data',
+        fileType: 'tanf',
+        year: '2024',
+        quarter: 'Q1',
+      }
+
+      const storeState = {
+        ...initialState,
+        reports: {
+          submittedFiles: [uploadedFile],
+        },
+      }
+
+      mockExecuteSubmission.mockImplementation(async (fn) => {
+        await fn()
+      })
+
+      const { getByText } = renderComponent(storeState)
+
+      const submitButton = getByText('Submit Data Files')
+      fireEvent.click(submitButton)
+
+      await waitFor(() => {
+        expect(mockExecuteSubmission).toHaveBeenCalled()
+      })
+    })
   })
 
   describe('Cancel Functionality', () => {
     it('shows error modal when cancel is clicked with uploaded files', async () => {
       const uploadedFile = {
         fileName: 'test.txt',
-        section: 'Quarter 1 (October - December)',
-        fileType: 'pia',
+        section: 'Active Case Data',
+        fileType: 'tanf',
         year: '2024',
+        quarter: 'Q1',
       }
 
       const storeState = {
@@ -382,7 +490,6 @@ describe('QuarterFileUploadForm', () => {
       fireEvent.click(cancelButton)
 
       // The error modal visibility is controlled by ReportsContext
-      // We can verify the state change through the context
       await waitFor(() => {
         expect(cancelButton).toBeInTheDocument()
       })
@@ -449,16 +556,23 @@ describe('QuarterFileUploadForm', () => {
         expect(alert).toBeInTheDocument()
       })
     })
+
+    it('does not scroll when alert is not active', () => {
+      renderComponent()
+
+      expect(window.HTMLElement.prototype.scrollIntoView).not.toHaveBeenCalled()
+    })
   })
 
-  describe('Quarter Formatting', () => {
-    it('formats single quarter correctly', () => {
+  describe('Section Formatting', () => {
+    it('maps section indices correctly', () => {
       const uploadedFiles = [
         {
-          fileName: 'q1.txt',
-          section: 'Quarter 1 (October - December)',
-          fileType: 'pia',
+          fileName: 'test1.txt',
+          section: 'Active Case Data',
+          fileType: 'tanf',
           year: '2024',
+          quarter: 'Q1',
         },
       ]
 
@@ -473,19 +587,28 @@ describe('QuarterFileUploadForm', () => {
       // Component renders without errors
     })
 
-    it('formats multiple quarters with "and" correctly', () => {
+    it('handles multiple sections with proper formatting', () => {
       const uploadedFiles = [
         {
-          fileName: 'q1.txt',
-          section: 'Quarter 1 (October - December)',
-          fileType: 'pia',
+          fileName: 'test1.txt',
+          section: 'Active Case Data',
+          fileType: 'tanf',
           year: '2024',
+          quarter: 'Q1',
         },
         {
-          fileName: 'q2.txt',
-          section: 'Quarter 2 (January - March)',
-          fileType: 'pia',
+          fileName: 'test2.txt',
+          section: 'Aggregate Data',
+          fileType: 'tanf',
           year: '2024',
+          quarter: 'Q1',
+        },
+        {
+          fileName: 'test3.txt',
+          section: 'Stratum Data',
+          fileType: 'tanf',
+          year: '2024',
+          quarter: 'Q1',
         },
       ]
 
@@ -505,26 +628,29 @@ describe('QuarterFileUploadForm', () => {
     it('passes correct props to FileUpload components', () => {
       const { getByTestId } = renderComponent()
 
-      const q1Upload = getByTestId('file-upload-Quarter 1 (October - December)')
-      const input = q1Upload.querySelector('input')
+      const activeUpload = getByTestId('file-upload-Active Case Data')
+      const input = activeUpload.querySelector('input')
 
-      expect(input).toHaveAttribute('data-section', 'Quarter 1 (October - December)')
-      expect(input).toHaveAttribute('data-quarter', '1')
+      expect(input).toHaveAttribute('data-section', 'Active Case Data')
+      expect(input).toHaveAttribute('data-filetype', 'tanf')
     })
 
-    it('renders FileUpload for each quarter in correct order', () => {
+    it('renders FileUpload for each section in correct order', () => {
       const { getByTestId } = renderComponent()
 
-      const uploads = [
-        'file-upload-Quarter 1 (October - December)',
-        'file-upload-Quarter 2 (January - March)',
-        'file-upload-Quarter 3 (April - June)',
-        'file-upload-Quarter 4 (July - September)',
-      ]
-
-      uploads.forEach((testId) => {
-        expect(getByTestId(testId)).toBeInTheDocument()
+      fileUploadSections.forEach((section) => {
+        expect(getByTestId(`file-upload-${section}`)).toBeInTheDocument()
       })
+    })
+
+    it('passes year and quarter from context to FileUpload', () => {
+      const { getByTestId } = renderComponent()
+
+      const activeUpload = getByTestId('file-upload-Active Case Data')
+      const input = activeUpload.querySelector('input')
+
+      expect(input).toHaveAttribute('data-year', '')
+      expect(input).toHaveAttribute('data-quarter', '')
     })
   })
 
@@ -532,9 +658,10 @@ describe('QuarterFileUploadForm', () => {
     it('uses stt.id when stt is provided', async () => {
       const uploadedFile = {
         fileName: 'test.txt',
-        section: 'Quarter 1 (October - December)',
-        fileType: 'pia',
+        section: 'Active Case Data',
+        fileType: 'tanf',
         year: '2024',
+        quarter: 'Q1',
       }
 
       const storeState = {
@@ -544,7 +671,7 @@ describe('QuarterFileUploadForm', () => {
         },
       }
 
-      const stt = { id: 42, name: 'Test State' }
+      const stt = { id: 42, name: 'Test State', num_sections: 4 }
 
       mockExecuteSubmission.mockImplementation(async (fn) => {
         await fn()
@@ -563,9 +690,10 @@ describe('QuarterFileUploadForm', () => {
     it('handles undefined stt gracefully', async () => {
       const uploadedFile = {
         fileName: 'test.txt',
-        section: 'Quarter 1 (October - December)',
-        fileType: 'pia',
+        section: 'Active Case Data',
+        fileType: 'tanf',
         year: '2024',
+        quarter: 'Q1',
       }
 
       const storeState = {
@@ -587,6 +715,57 @@ describe('QuarterFileUploadForm', () => {
       await waitFor(() => {
         expect(mockExecuteSubmission).toHaveBeenCalled()
       })
+    })
+
+    it('renders only num_sections when specified', () => {
+      const stt = { id: 1, num_sections: 3 }
+      const { getByTestId, queryByTestId } = renderComponent(initialState, stt)
+
+      expect(getByTestId('file-upload-Active Case Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Closed Case Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Aggregate Data')).toBeInTheDocument()
+      expect(queryByTestId('file-upload-Stratum Data')).not.toBeInTheDocument()
+    })
+
+    it('renders only 1 section when num_sections is 1', () => {
+      const stt = { id: 1, num_sections: 1 }
+      const { getByTestId, queryByTestId } = renderComponent(initialState, stt)
+
+      expect(getByTestId('file-upload-Active Case Data')).toBeInTheDocument()
+      expect(
+        queryByTestId('file-upload-Closed Case Data')
+      ).not.toBeInTheDocument()
+      expect(
+        queryByTestId('file-upload-Aggregate Data')
+      ).not.toBeInTheDocument()
+      expect(queryByTestId('file-upload-Stratum Data')).not.toBeInTheDocument()
+    })
+
+    it('uses default 4 sections when stt has no num_sections property', () => {
+      const stt = { id: 1 } // No num_sections property
+      const { getByTestId } = renderComponent(initialState, stt)
+
+      expect(getByTestId('file-upload-Active Case Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Closed Case Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Aggregate Data')).toBeInTheDocument()
+      expect(getByTestId('file-upload-Stratum Data')).toBeInTheDocument()
+    })
+
+    it('uses stt.num_sections when it is 0 (falsy but defined)', () => {
+      const stt = { id: 1, num_sections: 0 }
+      const { queryByTestId } = renderComponent(initialState, stt)
+
+      // Should render 0 sections (slice(0, 0) returns empty array)
+      expect(
+        queryByTestId('file-upload-Active Case Data')
+      ).not.toBeInTheDocument()
+      expect(
+        queryByTestId('file-upload-Closed Case Data')
+      ).not.toBeInTheDocument()
+      expect(
+        queryByTestId('file-upload-Aggregate Data')
+      ).not.toBeInTheDocument()
+      expect(queryByTestId('file-upload-Stratum Data')).not.toBeInTheDocument()
     })
   })
 
@@ -615,13 +794,14 @@ describe('QuarterFileUploadForm', () => {
       expect(getByText('Submit Data Files')).toBeInTheDocument()
     })
 
-    it('handles files with invalid section names', () => {
+    it('handles files with sections not in fileUploadSections', () => {
       const uploadedFiles = [
         {
           fileName: 'test.txt',
-          section: 'Invalid Section',
-          fileType: 'pia',
+          section: 'Unknown Section',
+          fileType: 'tanf',
           year: '2024',
+          quarter: 'Q1',
         },
       ]
 
@@ -651,50 +831,39 @@ describe('QuarterFileUploadForm', () => {
 
       await waitFor(() => {
         const alert = getByRole('alert')
-        expect(alert).toHaveTextContent('No changes have been made to data files')
+        expect(alert).toHaveTextContent(
+          'No changes have been made to data files'
+        )
       })
 
       expect(mockExecuteSubmission).not.toHaveBeenCalled()
     })
+
+    it('handles zero num_sections', () => {
+      const stt = { id: 1, num_sections: 0 }
+      const { queryByTestId } = renderComponent(initialState, stt)
+
+      expect(
+        queryByTestId('file-upload-Active Case Data')
+      ).not.toBeInTheDocument()
+      expect(
+        queryByTestId('file-upload-Closed Case Data')
+      ).not.toBeInTheDocument()
+      expect(
+        queryByTestId('file-upload-Aggregate Data')
+      ).not.toBeInTheDocument()
+      expect(queryByTestId('file-upload-Stratum Data')).not.toBeInTheDocument()
+    })
   })
 
-  describe('Console Logging', () => {
-    it('logs submission process', async () => {
-      const consoleSpy = jest.spyOn(console, 'log').mockImplementation()
+  describe('File Type Handling', () => {
+    it('displays correct file type in labels for TANF', () => {
+      const { getByText } = renderComponent()
 
-      const uploadedFile = {
-        fileName: 'test.txt',
-        section: 'Quarter 1 (October - December)',
-        fileType: 'pia',
-        year: '2024',
-      }
-
-      const storeState = {
-        ...initialState,
-        reports: {
-          submittedFiles: [uploadedFile],
-        },
-      }
-
-      mockExecuteSubmission.mockImplementation(async (fn) => {
-        await fn()
-      })
-
-      const { getByText } = renderComponent(storeState)
-
-      const submitButton = getByText('Submit Data Files')
-      fireEvent.click(submitButton)
-
-      await waitFor(() => {
-        expect(consoleSpy).toHaveBeenCalledWith('onSubmit called')
-      })
-
-      consoleSpy.mockRestore()
+      expect(getByText('Section 1 TANF - Active Case Data')).toBeInTheDocument()
     })
 
-    it('logs error when no files uploaded', async () => {
-      const consoleSpy = jest.spyOn(console, 'log').mockImplementation()
-
+    it('displays correct file type in labels for SSP-MOE', () => {
       const storeState = {
         ...initialState,
         reports: {
@@ -704,14 +873,41 @@ describe('QuarterFileUploadForm', () => {
 
       const { getByText } = renderComponent(storeState)
 
-      const submitButton = getByText('Submit Data Files')
-      fireEvent.click(submitButton)
+      // Note: The file type comes from ReportsContext, which defaults to 'tanf'
+      // In a real scenario, the context would need to be set to 'ssp-moe'
+      expect(getByText('Section 1 TANF - Active Case Data')).toBeInTheDocument()
+    })
+  })
 
-      await waitFor(() => {
-        expect(consoleSpy).toHaveBeenCalledWith('Setting error alert')
-      })
+  describe('Context Integration', () => {
+    it('uses yearInputValue from context', () => {
+      const { getByTestId } = renderComponent()
 
-      consoleSpy.mockRestore()
+      const activeUpload = getByTestId('file-upload-Active Case Data')
+      const input = activeUpload.querySelector('input')
+
+      // Year comes from ReportsContext, which defaults to ''
+      expect(input).toHaveAttribute('data-year', '')
+    })
+
+    it('uses quarterInputValue from context', () => {
+      const { getByTestId } = renderComponent()
+
+      const activeUpload = getByTestId('file-upload-Active Case Data')
+      const input = activeUpload.querySelector('input')
+
+      // Quarter comes from ReportsContext, which defaults to ''
+      expect(input).toHaveAttribute('data-quarter', '')
+    })
+
+    it('uses fileTypeInputValue from context', () => {
+      const { getByTestId } = renderComponent()
+
+      const activeUpload = getByTestId('file-upload-Active Case Data')
+      const input = activeUpload.querySelector('input')
+
+      // File type comes from ReportsContext, which defaults to 'tanf'
+      expect(input).toHaveAttribute('data-filetype', 'tanf')
     })
   })
 })
